Create YoutubeFeedTest.html
This is a prototype of a Youtube Feed which:
1. Pulls only the latest video from each channel in a list.
2. Displays each video in reverse chrono order.

Similar to Royal Road Follow List feed.

Easy to expand; just modify channelIds. 
Only hassle is that you have to get the channel Ids from the source of the channel page.
I wonder if the json export from something like newPipe includes channel Ids for convenience?